Fix infinite hang when loading a PNG
As it turns out we cannot just busy wait on a UnityWebRequest to
complete, that will hang forever.

This changes the PNG loader to only go through UnityWebRequest if the
hint is BatchAsync or above. If the unity web request is not done then
it will fall back to a regular file read and LoadRawData call. This
happens even if we have started a web request.