Avoid unloading textures while their asset bundle is still loaded
As it turns out if you destroy a texture loaded from an asset bundle
while that asset bundle is still live then it actually completely
deletes the texture object within the bundle.

To avoid this being an issue we instead hold onto a reference to the
texture until the asset bundle is unloaded.