Use AllocHGlobal for persistent allocations
From the profiler, it looks like Persistent allocations actually get
freed on the main thread even when disposed of in a job. With large
allocations this can actually be quite slow.

This changes the internals to use Marshal.AllocHGlobal to allocate
native arrays. AllocHGlobal can be freed in a job, so this avoids
slowdown on the main thread.