Copy data into the texture in the background for async loads
Calling LoadRawTextureData can actually take a while. This makes sense
in sync modes, since we really only get one chance to launch async work,
but for async modes we can afford to spend an extra frame doing the copy
in the background.

For slower machines this appears to make a pretty big difference.